+++ src/main/java/ru/mt/domain/AccountBalanceCallResult.java
@@ -5,6 +5,8 @@ import lombok.Getter;
 import lombok.ToString;
 import ru.mt.utils.TimeUtils;
 
+import java.math.BigDecimal;
+
 @Builder
 @Getter
 @ToString
@@ -12,7 +14,7 @@ public class AccountBalanceCallResult {
     private final long ts = TimeUtils.getTimestamp();
 
     private final String callId;
-    private final double amount;
+    private final BigDecimal amount;
     private final ReservationStatus reservationStatus;
     private final String errorMessage;
